Basic Design UI: Part 1 (#471)
This change is part 1 of adding a very simple banner design UI. It adds client
side routing etc and a simple list of the available designs by name (which
requires making a request to the API).

Navigation to the UI will be hidden in code/production, but can be navigated to
directly (for developers).

This includes a rename of the route param from testName to name. This param
isn't always the name of a test and I'm about to add another use case where it
isn't, instead it's the name of a design. So let's change the language to
something more generic which fits all use cases.