[TS only] banner variants can have a template OR a design
I've extended the type of the template property of a banner variant to
be a BannerUI. This is _either_ a BannerTemplate OR a BannerDesignName
which has the shape { name: string }. I think this will be backwards
compatible with the existing shape of data while allowing us to extend
with this new concept without mixing up the two. Ideally I'd have
probably had something like:

type BannerUI = { designName: string } | { componentName: string }

but that would require migrating data. Perhaps one for the future.