Support for 3 different image sizes (#475)
Adds support for multiple image URLs for different screen sizes: mobile,
tablet/desktop and wide. From talking to Rik the convention is that 3 images
are specified (although in the component 6 are supported, we specify 3 pairs).

I've slightly re-structured the BannerDesign type to have a complex "image"
property which contains the 3 image URLs plus the alt text.