Banner Design UI: Part 5 (Banner variants have a template or a design) (#476)
This PR updates the banner variant model to have either a template (as it does
now) or a banner design. When creating/editing a banner variant you will be
able to choose between selecting a variant or selecting a design.

This toggling is hidden behind a feature flag in CODE and PROD, so it isn't
actually possible to give a banner variant a design in these environments. This
is a good thing because currently SDC wouldn't know what to do in the case
where a banner variant has a design, so would break.

The banner variant model has been extended in a backwards compatible way which
I think means this change is safe to deploy. The TS variant model has been
updated to go from having this:

template: BannerTemplate

to having this:

template: BannerTemplate | BannerDesignName

where BannerDesignName looks like this:

export interface BannerDesignName {
  designName: string;
}

So for all existing cases, where a BannerTemplate is specified, things will
continue to work as they are.

On the server when creating/updating variants there isn't any validation
currently that the banner design specified actually exists. I'm intending to
add this in a future PR.

Note: previewing doesn't currently work for variants with a design (and won't
work until SDC has been updated).