First designable banner colour fields (#499)
This PR adds the first designable banner colour fields: body text and
background colour.

SDC counterpart: guardian/support-dotcom-components#966

Colours are entered as hex strings, validated and parsed into hex colour
objects before sending to the server.

For now the individual r, g and b fields on the object are typed as strings in
TS and Scala. In the SDC PR I was able to use string template types to be more
restrictive but ran into eslint issues here (eslint needs upgrading to support
this TS feature) and I didn't want to block this work.