Allow removing the choice card button bg colour
This affects all optional colour fields, which was previously slightly
broken.

The field isn't required, but previously unsetting it was broken
because we always tried to call stringToHexColour on the value, which
throws an error for an invalid colour string. This is supported by a new
OptionalColourInput component to complement the existing ColourInput
component which is now the component to use when colour is required.

This is implement entirely using the original ColourInput component
which is now generic. The "required" prop is no longer available on the
exported components.