Support for URLs with query string args in link tracking builder
Before, the link tracking builder tool did not allow any existing query
string args in the URL specified. This causes a problem for using the
tool with the generic checkout on the support site which requires
product and ratePlan as query string args. So instead show an error if
there are any tracking params in the URL specified, but allow other
query string args.